Refactor for easier diffing
